Replace configuration files to tie things together a bit better
Passing the "build" option to the script no longer configures MariaDB.
Configuration can be done with the new "setupconfig" option, which runs
the MariaDB secure utility (set password, remove anon users, etc) and
downloads a new set of configuration files (including those for PHP) so
that the server is in a more "workable" state after the initial
download and setup.

"nginx_standard" has been removed as well.  It doesn't take that much
effort to edit the script and remove what you don't want, and I'm not
going to be updating multiple copies of this script when versions
change, etc.